Got expand-state saving working in the profiles controller, in addition to a lot of random code stuff that Sam would not approve of having in the same commit.

+++ source/infrastructure/J3ByteSource.h
@@ -6,7 +6,7 @@
 
 @protocol J3ByteSource
 
-- (unsigned) availableBytes;
+- (NSUInteger) availableBytes;
 - (BOOL) hasDataAvailable;
 - (void) poll;
 - (NSData *) readExactlyLength: (size_t) length;
